Add files via upload
This is a Jupyter notebook showing the EDA done on the data including data cleaning and visualization. EDA is a very important step for any Data Science project. We get to see how the data looks, what insights can be generated and what variables are to be considered for the next steps i.e modeling and deployment.